requirements.txt
requests==2.31.0
gitpython==3.1.40
openai==1.3.0
markdown==3.5.1
#weasyprint==61.0
pygments==2.16.1
networkx==3.2.1
plotly==5.17.0
python-dotenv==1.0.0
reportlab==4.4.3
numpy==2.3.3